Refactor settings context into base context and improve type safety
