+++ src/components/settings/contexts/SettingsContext.tsx
@@ -1,17 +1,11 @@
-import React, { createContext, useContext, useReducer, useCallback, useEffect, useRef, useState, useMemo } from 'react';
+import React, { useReducer, useCallback, useEffect, useRef, useState, useMemo } from 'react';
 import { doc, getDoc, Timestamp } from 'firebase/firestore';
 import { useAuth } from '../../../contexts/AuthContext';
 import { db } from '../../../lib/firebase';
 import StorageService from '../../../lib/storage';
 import logger from '../../../lib/logger';
-import type {
-  DataSettings,
-  NotificationPreferences,
-  PrivacySettings,
-  SettingsAction,
-  SettingsState,
-  UserProfile,
-} from '../types';
+import type { DataSettings, NotificationPreferences, PrivacySettings, SettingsAction, SettingsState, UserProfile } from '../types';
+import { SettingsContext, type SettingsContextType } from './SettingsContextBase';
 
 // 기본 설정값
 const getDefaultSettings = (user: { displayName?: string; email?: string; photoURL?: string } | null): SettingsState => {
@@ -185,36 +179,7 @@ function settingsReducer(
   }
 }
 
-interface SettingsContextType {
-  settings: SettingsState;
-  updateSettings: (action: SettingsAction) => void;
-  updateProfile: (updates: Partial<UserProfile>) => void;
-  updateNotifications: (updates: Partial<NotificationPreferences>) => void;
-  updatePrivacy: (updates: Partial<PrivacySettings>) => void;
-  updateData: (updates: Partial<DataSettings>) => void;
-  saveSettings: () => Promise<void>;
-  loadSettings: () => Promise<void>;
-  resetToDefaults: () => void;
-  uploadAvatar: (
-    file: File,
-    onProgress?: (progress: number) => void
-  ) => Promise<string>;
-  deleteAvatar: () => Promise<void>;
-  loading: boolean;
-  saving: boolean;
-  uploadingAvatar: boolean;
-  error: string | null;
-}
-
-const SettingsContext = createContext<SettingsContextType | undefined>(undefined);
-
-export const useSettingsContext = () => {
-  const context = useContext(SettingsContext);
-  if (!context) {
-    throw new Error('useSettingsContext must be used within a SettingsProvider');
-  }
-  return context;
-};
+// Context and hook moved to SettingsContextBase to satisfy react-refresh rule
 
 interface SettingsProviderProps {
   children: React.ReactNode;
@@ -241,7 +206,7 @@ export const SettingsProvider: React.FC<SettingsProviderProps> = ({ children }) 
       logger.error('SettingsContext', 'Error accessing auth context', error);
       return null;
     }
-  }, [authContext.user?.uid]);
+  }, [authContext.user]);
 
   const STORAGE_KEY = 'moonwave-settings' as const;
 
@@ -374,7 +339,7 @@ export const SettingsProvider: React.FC<SettingsProviderProps> = ({ children }) 
     } finally {
       setLoading(false);
     }
-  }, [currentUser?.uid, authContext.user]);
+  }, [currentUser, authContext.user]);
 
   const loadSettings = useCallback(async () => {
     await loadSettingsInternal();
@@ -589,7 +554,7 @@ export const SettingsProvider: React.FC<SettingsProviderProps> = ({ children }) 
       isInitializedRef.current = true;
       loadSettingsInternal();
     }
-  }, [currentUser?.uid]);
+  }, [currentUser?.uid, loadSettingsInternal]);
 
   useEffect(() => {
     if (currentUser?.uid) {
@@ -605,7 +570,7 @@ export const SettingsProvider: React.FC<SettingsProviderProps> = ({ children }) 
         }
       };
     }
-  }, [currentUser?.uid]);
+  }, [currentUser?.uid, loadSettingsInternal]);
 
   const updateProfile = useCallback((updates: Partial<UserProfile>) => {
     dispatch({ type: 'UPDATE_PROFILE', payload: updates });
